fixed seg fault

+++ normalise.c
@@ -6,7 +6,7 @@
 /*   By: rmakoni <rmakoni@student.42heilbronn.de    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/10 17:00:34 by rmakoni           #+#    #+#             */
-/*   Updated: 2024/12/16 12:21:35 by rmakoni          ###   ########.fr       */
+/*   Updated: 2024/12/16 18:55:14 by rmakoni          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ int	has_duplicates(t_list *stack)
 		checker = current->next;
 		while (checker != NULL)
 		{
-			if ((int)current->content == (int)checker->content)
+			if ((int)(intptr_t)current->content == (int)(intptr_t)checker->content)
 			{
 				return (1);
 			}
@@ -41,11 +41,11 @@ int	get_largest(t_list *stack)
 	largest = stack;
 	while (stack != NULL)
 	{
-		if ((int)largest->content < (int)stack->content)
+		if ((int)(intptr_t)largest->content < (int)(intptr_t)stack->content)
 			largest = stack;
 		stack = stack->next;
 	}
-	return ((int)largest->content);
+	return ((int)(intptr_t)largest->content);
 }
 
 int	get_smallest(t_list *stack)
@@ -55,11 +55,11 @@ int	get_smallest(t_list *stack)
 	smallest = stack;
 	while (stack != NULL)
 	{
-		if ((int)smallest->content > (int)stack->content)
+		if ((int)(intptr_t)smallest->content > (int)(intptr_t)stack->content)
 			smallest = stack;
 		stack = stack->next;
 	}
-	return ((int)smallest->content);
+	return ((int)(intptr_t)smallest->content);
 }
 void	normalise(t_list **stack)
 {
@@ -81,7 +81,7 @@ void	normalise(t_list **stack)
 	current = *stack;
 	while (current != NULL)
 	{
-		i_current = (int)current->content;
+		i_current = (int)(intptr_t)current->content;
 		current->content = (void *)(intptr_t)(k * ((i_current - smallest)
 					/ (largest - smallest)));
 		current = current->next;
